move pgp signing to repository worker
