+++ lib/workers/global/index.js
@@ -5,7 +5,7 @@ const repositoryWorker = require('../repository');
 const cache = require('./cache');
 const { appName } = require('../../config/app-strings');
 const { autodiscoverRepositories } = require('./autodiscover');
-const { setMeta, setGitPrivateKey } = require('./meta');
+const { setMeta } = require('./meta');
 
 module.exports = {
   start,
@@ -18,7 +18,6 @@ async function start() {
   try {
     let config = await configParser.parseConfigs(process.env, process.argv);
     setMeta(config);
-    await setGitPrivateKey(config);
     config = await autodiscoverRepositories(config);
     if (config.repositories.length === 0) {
       logger.warn(
