+++ lib/workers/global/meta.js
@@ -1,9 +1,7 @@
 const addrs = require('email-addresses');
-const Git = require('simple-git/promise');
 
 module.exports = {
   setMeta,
-  setGitPrivateKey,
 };
 
 function setMeta(config) {
@@ -25,21 +23,5 @@ function setMeta(config) {
       name: gitAuthorParsed.name,
       email: gitAuthorParsed.address,
     };
-  }
-}
-async function setGitPrivateKey(config) {
-  if (config.gitPrivateKey && config.localDir) {
-    try {
-      const cwd = config.localDir;
-      const git = Git(cwd).silent(true);
-      await git.raw([
-        'config',
-        '--global',
-        'user.signingkey',
-        config.gitPrivateKey,
-      ]);
-    } catch (err) {
-      logger.warn({ err }, 'Error in setting git private key to author');
-    }
   }
 }
